`buy_ticket` WIP
