+++ bot/handlers/wheel_of_fortune_handlers/__init__.py
@@ -1,3 +1,3 @@
 from .wheel_of_fortune import router as wheel_of_fortune_router
-
-routers = [wheel_of_fortune_router]
+from .buy_ticket import router as buy_ticket_router
+routers = [wheel_of_fortune_router, buy_ticket_router]
